chore: refactor rules & traversing of cmds
