+++ validator/validator.go
@@ -1,6 +1,7 @@
 package validator
 
 import (
+	"github.com/aerogear/charmil/validator/rules"
 	"github.com/spf13/cobra"
 )
 
@@ -10,5 +11,7 @@ type Rule interface {
 	// Validate validates the cobra command
 	// and returns errors according to the provided rules
 	// verbose option is present which prints debug info
-	Validate(cmd *cobra.Command, verbose bool) []Error
+	// Validate(cmd *cobra.Command, verbose bool) []rules.Error
+	ValidateLength(cmd *cobra.Command, verbose bool) []rules.ValidationError
+	ValidateMustPresent(cmd *cobra.Command, verbose bool) []rules.ValidationError
 }
